booleans in argument parsing have issues:
When you use `bool` as the type for an argument in `argparse.ArgumentParser`, it doesn't behave as you might expect. This is because `argparse` treats the argument as a string and then tries to convert it to a boolean. In Python, `bool("False")` and `bool("True")` both return `True` because non-empty strings are truthy.

So, if you pass `--normalize_features False` in the command line, `argparse` will still interpret it as `True` because it's a non-empty string.

Instead, you can use `argparse`'s `store_true` or `store_false` actions to handle boolean values. For example:

```python
self.parser.add_argument(
    "--normalize_features",
    action='store_true',
    help="Whether to normalize the edge features in meshcnn meshes using std and mean",
)
```

In this case, including the `--normalize_features` flag without any argument will set `normalize_features` to `True`. If the flag is not included, `normalize_features` will be `False`.